Separate functional tests from E2E tests
