+++ .vscode/launch.json
@@ -8,10 +8,22 @@
       "type": "node",
       "request": "launch",
       "name": "Debug Unit Jest Tests",
-      "program": "${workspaceRoot}/node_modules/jest/bin/jest.js",
+      "runtimeExecutable": "npm",
       "cwd": "${workspaceFolder}",
-      "runtimeArgs": ["--nolazy", "--inspect"],
-      "args": ["--runInBand", "--watch", "--config", "${workspaceRoot}/jest.config.js", "${fileBasenameNoExtension}"],
+      "runtimeArgs": ["run-script", "test:unit"],
+      "args": ["--", "--runInBand", "--watch", "${fileBasenameNoExtension}"],
+      "sourceMaps": true,
+      "outputCapture": "std",
+      "console": "integratedTerminal"
+    },
+    {
+      "type": "node",
+      "request": "launch",
+      "name": "Debug Functional Jest Tests",
+      "runtimeExecutable": "npm",
+      "cwd": "${workspaceFolder}",
+      "runtimeArgs": ["run-script", "test:func"],
+      "args": ["--", "--runInBand", "--watch", "${fileBasenameNoExtension}"],
       "sourceMaps": true,
       "outputCapture": "std",
       "console": "integratedTerminal"
@@ -20,10 +32,10 @@
       "type": "node",
       "request": "launch",
       "name": "Debug E2E Jest Tests",
-      "program": "${workspaceRoot}/node_modules/jest/bin/jest.js",
+      "runtimeExecutable": "npm",
       "cwd": "${workspaceFolder}",
-      "runtimeArgs": ["--nolazy", "--inspect"],
-      "args": ["--runInBand", "--watch", "--config", "${workspaceRoot}/jest.e2e.js", "${fileBasenameNoExtension}"],
+      "runtimeArgs": ["run-script", "test:e2e"],
+      "args": ["--", "--runInBand", "--watch", "${fileBasenameNoExtension}"],
       "sourceMaps": true,
       "outputCapture": "std",
       "console": "integratedTerminal"
