+++ jest.func.js
@@ -1,4 +1,4 @@
 module.exports = {
   ...require('./jest.config'),
-  testMatch: ['**/__e2e__/**/*.test.ts'],
+  testMatch: ['**/__functional__/**/*.test.ts'],
 };
