+++ src/__e2e__/README.md
@@ -1,0 +1,1 @@
+These tests are true E2E tests covering major beachball scenarios. Tests that cover specific helper functions but need filesystem-based fixtures should go in `__functional__` instead.
