+++ src/__functional__/monorepo/getPackageInfos.test.ts
