+++ src/__functional__/monorepo/getScopedPackages.test.ts
